+++ lib/eventsource/eventsource-stream.js
@@ -71,7 +71,7 @@ class EventSourceStream extends Transform {
   constructor (options = {}) {
     options.readableObjectMode = true
     super(options)
-    this.state = options.eventSourceState
+    this.state = options.eventSourceState || {}
     if (options.push) {
       this.push = options.push
     }
